refactor: moved ParaChestBlockEntity to content/block/entity/ with the rest of the BlockEntity classes
