+++ src/main/java/dev/galacticraft/mod/content/GCBlockEntityTypes.java
@@ -29,7 +29,6 @@ import dev.galacticraft.mod.content.block.entity.networked.FluidPipeWalkwayBlock
 import dev.galacticraft.mod.content.block.entity.networked.GlassFluidPipeBlockEntity;
 import dev.galacticraft.mod.content.block.entity.networked.WireBlockEntity;
 import dev.galacticraft.mod.content.block.entity.networked.WireWalkwayBlockEntity;
-import dev.galacticraft.mod.content.block.special.ParaChestBlockEntity;
 import dev.galacticraft.mod.content.block.special.launchpad.LaunchPadBlockEntity;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
