+++ src/main/java/dev/galacticraft/mod/content/block/entity/ParaChestBlockEntity.java
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package dev.galacticraft.mod.content.block.special;
+package dev.galacticraft.mod.content.block.entity;
 
 import dev.galacticraft.mod.content.GCBlockEntityTypes;
 import dev.galacticraft.mod.content.entity.ScalableFuelLevel;
